🔨 Pull reusable functions out of portal flow
