+++ util/urlFor.js
@@ -1,0 +1,3 @@
+module.exports = function urlFor(baseUrl, dataset) {
+  return `${baseUrl}/mbaas/sync/${dataset}`;
+};
